pt_sweep: int() the token budget so fractional chinchillas stay integral
N_TOKENS = BASE_TOKENS * CHINCHILLA is a float at chinchilla 0.25 / 0.5, and
warmup_steps derived from it too (600309760.0 / 57.0 at 60M c0.5). The values
match the main matrix; only the type was wrong, and it would have leaked into
the generated YAML. Same guard as pretraining_matrix._tokens_for.

+++ launch_jolmo/pt_sweep_60m_chin4.py
@@ -143,7 +143,10 @@ DEFAULT_WEIGHT_DECAY = 0.1              # matches SHARED_MODEL_PARAMS in the mai
 # dict holds the 0.06B entry only, so a retargeted size raised KeyError. 0.06B
 # still resolves to the same 20 x 60,030,976 through the override inside it.
 BASE_TOKENS = _base_tokens_for(MODEL_TYPE)
-N_TOKENS = BASE_TOKENS * CHINCHILLA
+# int(): a fractional chinchilla (0.25, 0.5) otherwise makes n_tokens and the
+# derived warmup_steps floats, which leak into the generated YAML. Same guard
+# pretraining_matrix._tokens_for applies; the values are unchanged.
+N_TOKENS = int(BASE_TOKENS * CHINCHILLA)
 
 _SIZE, _PROFILE = active_profile()
